+++ app/sprinkles/account/locale/fr_FR/validate.php
@@ -16,6 +16,6 @@
 return [
     'VALIDATE' => [
         'PASSWORD_MISMATCH' => 'Votre mot de passe et votre mot de passe de confirmation doivent correspondre.',
-        // 'USERNAME'          => 'Username may consist only of lowercase letters, numbers, \'.\', \'-\', and \'_\'.',
+        'USERNAME'          => 'Le nom d\'utilisateur ne peut contenir que des lettres minuscules, des chiffres, \'.\', \'-\', et \'_\'.',
     ],
 ];