+++ app/sprinkles/account/locale/pt_BR/validate.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @copyright Copyright (c) 2019 Alexander Weissman
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/LICENSE.md (MIT License)
+ */
+
+/**
+ * Brazilian Portuguese message token translations for the 'account' sprinkle.
+ *
+ * @author Maxwell Kenned (kenned123@gmail.com)
+ */
+return [
+    'VALIDATE' => [
+        'PASSWORD_MISMATCH' => 'Confirmação de senha inválida.',
+        'USERNAME'          => 'O nome de usuário pode consistir apenas em letras minúsculas, numeros, \'.\', \'-\', e \'_\'.',
+    ],
+];