+++ app/sprinkles/admin/tests/Integration/Controller/UserControllerTest.php
@@ -796,7 +796,7 @@ class UserControllerTest extends TestCase
 
         // Set post data
         $data = [
-            'value' => 'bar',
+            'first_name' => 'bar',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -809,6 +809,41 @@ class UserControllerTest extends TestCase
         // Make sure user was update
         $editedUser = User::where('user_name', $user->user_name)->first();
         $this->assertSame('bar', $editedUser->first_name);
+        $this->assertNotSame($user->first_name, $editedUser->first_name);
+        $this->assertSame($user->last_name, $editedUser->last_name);
+
+        // Test message
+        /** @var \UserFrosting\Sprinkle\Core\Alert\AlertStream $ms */
+        $ms = $this->ci->alerts;
+        $messages = $ms->getAndClearMessages();
+        $this->assertSame('success', end($messages)['type']);
+    }
+
+    /**
+     * @depends testControllerConstructorWithUser
+     * @depends testUpdateField
+     * @param UserController $controller
+     */
+    public function testUpdateFieldWithDeprecatedSupport(UserController $controller)
+    {
+        // Create a user
+        $user = $this->createTestUser();
+
+        // Set post data
+        $data = [
+            'value' => 'deprecated', //<-- Use old `value`
+        ];
+        $request = $this->getRequest()->withParsedBody($data);
+
+        // Get controller stuff
+        $result = $controller->updateField($request, $this->getResponse(), ['user_name' => $user->user_name, 'field' => 'first_name']);
+        $this->assertSame($result->getStatusCode(), 200);
+        $this->assertJson((string) $result->getBody());
+        $this->assertSame('[]', (string) $result->getBody());
+
+        // Make sure user was update
+        $editedUser = User::where('user_name', $user->user_name)->first();
+        $this->assertSame('deprecated', $editedUser->first_name);
         $this->assertNotSame($user->first_name, $editedUser->first_name);
         $this->assertSame($user->last_name, $editedUser->last_name);
 
@@ -853,7 +888,7 @@ class UserControllerTest extends TestCase
 
         // Set post data
         $data = [
-            'value' => 'barbar',
+            'first_name' => 'barbar',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -889,7 +924,7 @@ class UserControllerTest extends TestCase
 
         // Set post data
         $data = [
-            'value' => $value,
+            'first_name' => $value,
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -907,7 +942,7 @@ class UserControllerTest extends TestCase
     {
         // Set post data
         $data = [
-            'value' => '1',
+            'flag_enabled' => '1',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -937,7 +972,7 @@ class UserControllerTest extends TestCase
     {
         // Set post data
         $data = [
-            'value' => '0',
+            'flag_enabled' => '0',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -967,7 +1002,7 @@ class UserControllerTest extends TestCase
     {
         // Set post data
         $data = [
-            'value' => '1',
+            'flag_verified' => '1',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -1000,7 +1035,7 @@ class UserControllerTest extends TestCase
 
         // Set post data
         $data = [
-            'value' => '0',
+            'flag_enabled' => '0',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -1018,7 +1053,7 @@ class UserControllerTest extends TestCase
     {
         // Set post data
         $data = [
-            'value' => '1234567890abc',
+            'password' => '1234567890abc',
         ];
         $request = $this->getRequest()->withParsedBody($data);
 
@@ -1056,7 +1091,7 @@ class UserControllerTest extends TestCase
 
         // Set post data
         $data = [
-            'value' => [['role_id' => $role->id]],
+            'roles' => [['role_id' => $role->id]],
         ];
         $request = $this->getRequest()->withParsedBody($data);